+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 1.4
 ---
 
+ * In `Shop\Service\SiteService` added `getCurrencies()` and `getCurrencySourceString()`
  * Ability to return formatted prices from product entities
  * Added `countAllStokes()` and `getAllStokes()` in `\Shop\Service\SiteService`
  * Multi-currencies support. Added `getConvertedPrice()` and `getConvertedStokePrice()` in product entity